MEI-819 Remove legacy build files
Now that cmake is used for mac and windows mordor builds
we do not need legacy project files.
In fact those files are misleading because they refer to old
version of openssl.

(Not for linux we retain the auto make files because
server team has not converted to cmake)

Change-Id: I1ef5101451731ac9cc3fd0a9fb23e07707f6ef1f
tethys-patch: 118315
Reviewed-on: https://gerrit.mozy.lab.emc.com/118316
Reviewed-by: Kevin Harris <Kevin.Harris@emc.com>